% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_category_matrix.R
\name{create_category_matrix}
\alias{create_category_matrix}
\title{Generates a matrix for lexicon Category results}
\usage{
create_category_matrix(lecat_result, inShiny = FALSE)
}
\arguments{
\item{lecat_result}{data frame output from the \link[lecat]{run_lecat_analysis} function}

\item{inShiny}{If inShiny is TRUE then shiny based notifications will be shown}
}
\value{
Passing the output of the \link[lecat]{run_lecat_analysis} function will return a data frame
}
\description{
Creates a file listing the total matches for each Category in the lexicon for each row in the corpus.
The file can be pasted as additional columns in the corpus, which allows for using filter and sorting functions to explore the data.
For example, to find all rows in an analysed corpus that match a specific lexicon category or combination of categories and how many times.
}
