% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cooccurrence_graph.R
\name{create_cooccurrence_graph}
\alias{create_cooccurrence_graph}
\title{Create a graph representation of the cooccurences of Categories}
\usage{
create_cooccurrence_graph(
  result,
  graph_filename = "result.graphml",
  cotable_filename = "cotable.csv",
  inShiny = FALSE
)
}
\arguments{
\item{result}{result from the \link[lecat]{run_lecat_analysis} function}

\item{graph_filename}{filename of the generated graph file}

\item{cotable_filename}{filename of the generated cooccurrence table}

\item{inShiny}{If inShiny is TRUE then shiny based notifications will be shown}
}
\value{
cooccurence table (At present.)
}
\description{
Calculates the cooccurrence of Categories.
A graph object is then created using the \link[igraph]{make_graph} function for
visualising in Gephi or similiar software.
}
\details{
Creates a cooccurrence graph
}
