% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cooccurrence_queries_graph.R
\name{create_cooccurrence_query_graph}
\alias{create_cooccurrence_query_graph}
\title{create_cooccurrence_query_graph}
\usage{
create_cooccurrence_query_graph(lecat_result, filename = NULL)
}
\arguments{
\item{lecat_result}{data frame output from the \link[lecat]{run_lecat_analysis} function}

\item{filename}{filename of the graphml graph file}
}
\value{
Returns a list containing a table of the co-occurrence of query terms
and a graph object when passed the output of the \link[lecat]{run_lecat_analysis} function. The nodes of the
undirected graph have the Type, Category and Column_examined attributes and edges are
weighted according to co-occurance. The graphi is written if a filename is specified.
}
\description{
Counts the co-occurrence of queries within each corpus element.
An graphml file is created and the co-occurrence table is returned.
Calculations are currently very inefficient and will be improved in the future.
}
