% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cotable.R
\name{create_cotable}
\alias{create_cotable}
\title{create_cotable}
\usage{
create_cotable(this_mat, that_mat)
}
\arguments{
\item{this_mat}{matrix of occurences of one item. Rows are items (Category, Type or Query) and columns are corpus elements}

\item{that_mat}{matrix of occurences of the other items. Rows and columns as this_mat.}
}
\value{
data frame containing node1, node2 and the cooccurrence frequency
}
\description{
Calculates a table of cooccurences
}
\details{
Generates a cooccurence table
}
