% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_unique_total_diagnostics.R
\name{create_unique_total_diagnostics}
\alias{create_unique_total_diagnostics}
\title{Generates diagnostics about the unique and total occurences of Types, Categories and Queries}
\usage{
create_unique_total_diagnostics(lecat_result, inShiny = FALSE)
}
\arguments{
\item{lecat_result}{data frame output from the \link[lecat]{run_lecat_analysis} function}

\item{inShiny}{If inShiny is TRUE then shiny based notifications will be shown}
}
\value{
Passing the output of the \link[lecat]{run_lecat_analysis} function will return a data frame with Type,
Category, Queries and Column_examined columns. In the output the unique and total occurances of Types, Category and Query
are reported in the format Term(total occurances, unique occurances).
}
\description{
Creates a file listing the total occurances of all queries and categories.
}
