% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_youtube_video_description.R
\name{download_youtube_video_description}
\alias{download_youtube_video_description}
\title{download_youtube_video_description}
\usage{
download_youtube_video_description(video_id, api_key, part = "snippet")
}
\arguments{
\item{video_id}{The id of the YouTube video}

\item{api_key}{An API key able to access the YouTube API v3}

\item{part}{The data you want returned. See the YouTube documentation for details. Defaults to 'snippet, statistics'.}
}
\value{
Data frame containing the descriptions and statistics of the YouTube video identified by the video_id parameter.
}
\description{
Downloads a single YouTube video title, ID, publish time and description.
}
