% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_youtube_video_descriptions.R
\name{download_youtube_video_descriptions}
\alias{download_youtube_video_descriptions}
\title{download_youtube_video_descriptions}
\usage{
download_youtube_video_descriptions(video_ids, api_key, part = "snippet",
  inShiny = FALSE)
}
\arguments{
\item{video_ids}{Character array containing multiple YouTube IDs}

\item{api_key}{An API key able to access the YouTube API v3}

\item{part}{The data you want returned. See the YouTube documentation for details. Defaults to 'snippet, statistics'}

\item{inShiny}{If inShiny is TRUE then shiny based notifications will be shown}
}
\value{
Data frame containing YouTube ID, title, published and description
}
\description{
Downloads the description of multiple videos using the YouTube V3 API.
Use the \link[lecat]{download_youtube_video_description} function for a single YouTube video.
}
\details{
Downloads the video descriptions of multiple YouTube videos using the V3 YouTube API.
}
