% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_video_ids.R
\name{extract_video_ids}
\alias{extract_video_ids}
\title{extract_video_ids}
\usage{
extract_video_ids(urls, resolve = TRUE, inShiny = FALSE)
}
\arguments{
\item{urls}{A character array of YouTube urls}

\item{resolve}{If resolve is TRUE then the function will attempt to resolve URLS which do not
appear to be of the format www.youtube.com/watch?v=xxxxxx}

\item{inShiny}{If inShiny is TRUE then notifications will be shown on the shiny app}
}
\value{
A character array of YouTube video IDs are returned.
These can then be passed to the comment or description download functions.
}
\description{
Extract YouTube IDs from URLs. The URL string is split and the segment contining
the video is extracted. Wrapper for the urltools function param_get.
}
