% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{run_app}
\usage{
run_app()
}
\description{
The lecat package contains several functions and the run_app
function starts a Shiny app for interacting with these function. The Shiny app is a
web based interface suitable for students with little experience of R to use in place
of calling the function themselves. The app can run either locally - started by the
run_app function - or externally by placing the package and app.R file on an external server.
}
\details{
Starts the Shiny App interface for the lecat package
}
