% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cooccurence_graph.R
\name{create_cooccurrence_graph}
\alias{create_cooccurrence_graph}
\title{create_cooccurence_graph}
\usage{
create_cooccurrence_graph(le_cat_result)
}
\arguments{
\item{le_cat_result}{result from the \link[lecat]{run_lecat_analysis} function}
}
\value{
cooccurence table (At present.)
}
\description{
Calculates the cooccurrence of either Types, Categories or Queries.
A graph object is then created using the \link[igraph]{make_graph} function for
visualising in Gephi or similiar software.
}
\details{
Creates a cooccurence graph
}
