% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cotable.R
\name{create_cotable}
\alias{create_cotable}
\title{create_cotable}
\usage{
create_cotable(this_mat, that_mat, this_mat_attr, that_mat_attr, level)
}
\arguments{
\item{this_mat}{matrix of occurences of one item. Rows are nodes - differs to either Type, Category or Query depending on the level passed to \link[lecat]{create_cooccurrence_graph} - and columns are corpus elements}

\item{that_mat}{matrix of occurences of the other items. Rows and columns as this_mat}

\item{this_mat_attr}{data frame containing a single nodes attributes. Attributes present - Type or Category -
depends on the level parameter passed to \link[lecat]{create_cooccurrence_graph}}

\item{that_mat_attr}{data frame containing attribute of all other nodes. Attributes differ as above depending on level}

\item{level}{level of cooccurrence analysis. Can be either 'Type', 'Category' or 'Query'.}
}
\value{
data frame containing node1, node2 and the cooccurrence frequency
}
\description{
Calculates a table of cooccurences of a single node to all other nodes. Called by \link[lecat]{create_cooccurrence_graph} function for each node in turn.
}
\details{
Generates a cooccurence table
}
