% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lecat_analysis.R
\name{run_lecat_analysis}
\alias{run_lecat_analysis}
\title{run_lecat_analysis}
\usage{
run_lecat_analysis(
  lexicon,
  corpus,
  searches,
  id = NaN,
  regex_expression = "\\\\Wquery\\\\W",
  inShiny = FALSE,
  case_sensitive = FALSE
)
}
\arguments{
\item{lexicon}{Lexicon dataframe as parsed by the \link[lecat]{parse_lexicon} function}

\item{corpus}{Corpus dataframe containing search columns present in the searches dataframe}

\item{searches}{Data frame with the columns 'Type' and 'Column'. Queries in each Type will be located in the corresponding corpus Column}

\item{id}{Column name to use for identifying differing corpus samples (e.g., YouTube video id). Autogenerated if no id is provided.}

\item{regex_expression}{Regex expression defining search. String defining the regex expression where the string 'query' will be replaced by the actual query term}

\item{inShiny}{If inShiny is TRUE then shiny based notifications will be shown}

\item{case_sensitive}{If case_sensitive is TRUE then the search will be case sensitive}
}
\value{
run_lecat_analysis returns a data frame containing the lexicon, the corresponding search column for the query type and the frequency of terms by corpus id
}
\description{
The occurance frequency of queries in a corpus is recorded in preperation for
futher analysis by the diagnostics or coocurance functions.
}
